[CMPT-5391] address PR feedback: SilenceUsage and extract resolveFilePath
- Add SilenceUsage: true to create, list, and update commands (consistent with del, get, lock, graph, version get/list)
- Extract duplicated resolveFilePath into cmd/pipeline/fileutil package
- Move TestResolveFilePath tests to fileutil package
