+++ cmd/pipeline/fileutil/fileutil_test.go
@@ -1,0 +1,48 @@
+// Copyright 2026 DataRobot, Inc. and its affiliates.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package fileutil
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+func TestResolveFilePath(t *testing.T) {
+	t.Run("positional only", func(t *testing.T) {
+		got, err := ResolveFilePath([]string{"a.py"}, "")
+		require.NoError(t, err)
+		assert.Equal(t, "a.py", got)
+	})
+
+	t.Run("flag only", func(t *testing.T) {
+		got, err := ResolveFilePath(nil, "b.py")
+		require.NoError(t, err)
+		assert.Equal(t, "b.py", got)
+	})
+
+	t.Run("both supplied", func(t *testing.T) {
+		_, err := ResolveFilePath([]string{"a.py"}, "b.py")
+		require.Error(t, err)
+		assert.Contains(t, err.Error(), "not both")
+	})
+
+	t.Run("neither supplied", func(t *testing.T) {
+		_, err := ResolveFilePath(nil, "")
+		require.Error(t, err)
+		assert.Contains(t, err.Error(), "required")
+	})
+}
